+++ tests/testthat/test-POLYFUN.R
@@ -1,5 +1,6 @@
 test_that("POLYFUN_SUSIE works", {
     
+    set.seed(1234)
     locus_dir <- file.path(tempdir(),echodata::locus_dir)
     dat <- echodata::BST1 
     LD_matrix <- echodata::BST1_LD_matrix
@@ -9,8 +10,8 @@ test_that("POLYFUN_SUSIE works", {
                           n_causal,
                           cs_snps){
         testthat::expect_equal(nrow(LD_matrix), nrow(daty))
-        testthat::expect_true(!all(c("POLYFUN_h2","CS","PP") %in% colnames(datx)))
-        testthat::expect_true(all(c("POLYFUN_h2","CS","PP") %in% colnames(daty)))
+        testthat::expect_true(!all(c("POLYFUN.h2","CS","PP") %in% colnames(datx)))
+        testthat::expect_true(all(c("POLYFUN.h2","CS","PP") %in% colnames(daty)))
         cs <- subset(daty, PP>.95)
         testthat::expect_equal(nrow(cs),n_causal) 
         testthat::expect_equal(cs[cs$CS==min(cs$CS),]$SNP,cs_snps)
@@ -27,7 +28,7 @@ test_that("POLYFUN_SUSIE works", {
               cs_snps="rs10003136") 
     
     #### POLYFUN: FINEMAP #####
-    dat3 <- echofinemap::POLYFUN(locus_dir=locus_dir,
+    dat3 <- echofinemap:: POLYFUN(locus_dir=locus_dir,
                                  dat=dat,
                                  LD_matrix = LD_matrix, 
                                  method = "FINEMAP",